New DelayLogForUpstream configuration option, and try submitblock(data) before getmemorypool(data)
Set to True if you want shares meeting the upstream target to wait for a response from the upstream server before logging them.
Otherwise, for such shares, upstreamResult will always be True and upstreamRejectReason will always be None.
Note that enabling this may cause shares to be logged out of order, or with the wrong timestamp (if your share logger uses the log-time rather than share-time).